refactor(cli): extract _shared.py leaf helpers from cli monolith
First slice of the cli monolith decomposition planned in
docs/refactor/cli-monolith-next-slice.md. Moves leaf helpers, option
dataclasses, and low-level constants (log-level map, progress bar
config, worker-count validation, hook path guards, version flag) from
phi_scan/cli/__init__.py into phi_scan/cli/_shared.py.

Private symbols are re-exported from phi_scan.cli via __all__ so
existing test imports continue to work unchanged, except for nine
monkeypatch targets on _GIT_DIR_PATH which now point at
phi_scan.cli._shared._GIT_DIR_PATH (the new home of the constant that
_reject_missing_git_directory actually reads). No behavior change.

cli/__init__.py shrinks 1492 → 1285 lines.
